Reorganise support directories #44
